requirements.txt
asgiref==3.4.1
Django==3.2.9
django-cors-headers==3.10.1
django-filter==21.1
django-resized==0.3.11
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
importlib-metadata==4.8.2
Markdown==3.3.6
Pillow==8.4.0
psycopg2-binary==2.9.2
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
whitenoise==5.3.0
zipp==3.6.0
